% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cov.R
\name{calc.cov}
\alias{calc.cov}
\title{Calculate the coefficient of variation}
\usage{
calc.cov(backslip_ref, offset_type, mean, stdev)
}
\arguments{
\item{backslip_ref}{A reference datatable containing confidence values.}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error.}

\item{mean}{A number referencing the column number of means in the reference
datatable 'backslip_ref'}

\item{stdev}{A number referencing the column number of standard deviations
in the reference datatable 'backslip_ref'}
}
\value{
The input datatable with an additional column of COV values
}
\description{
Calculate the coefficient of variation (COV) values for lateral, vertical,
or total offset values.
}
