% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_scale.R
\name{conf.scale}
\alias{conf.scale}
\title{Scale offset values by confidence}
\usage{
conf.scale(offset_data, backslip_ref, confidence_column, N)
}
\arguments{
\item{offset_data}{A datatable with x values as the first column and probability
distribution functions (y values) as subsequent columns.}

\item{backslip_ref}{A datatable containing confidence values.}

\item{confidence_column}{The column number in a 'backslip_ref' table with confidence
values.}

\item{N}{The integer number of offset sites
(i.e. # of files contained in input mat folder)}
}
\value{
A single datatable with backslip offsets scaled by measurement confidence
}
\description{
Multiply each probability distribution function contained in a data frame by
a confidence value. Make sure confidence values are >= 1. Values equal to 1
are highest confidence measurements. Larger values (>1) are lower confidence.
}
