% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{data.clip}
\alias{data.clip}
\title{Clip x extent of offset datatable}
\usage{
data.clip(data_table, offset_type, xmin, xmax)
}
\arguments{
\item{data_table}{A single table with x values as the first column and every
subsequent column is a PDF of y values for a single offset site.}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error.}

\item{xmin}{The minimum x value. Values can be negative or positive but must be less than xmax.}

\item{xmax}{The maximum x value. Values can be negative or positive but must be greater than xmin.}
}
\value{
A datatable similar to the input data table with a clipped extent for x values.
}
\description{
Clip the x bounds of offset datatable for better plotting.
}
