% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cov.R
\name{filter_covDAT}
\alias{filter_covDAT}
\title{Filter offset data by COV values}
\usage{
filter_covDAT(
  backslip_ref,
  offset_ID,
  offset_type,
  COV_values,
  COV_mat,
  backslip_data = backslip_data
)
}
\arguments{
\item{backslip_ref}{A reference datatable containing coefficient of variation (COV) values.}

\item{offset_ID}{The column number in the reference datatable 'backslip_ref' corresponding
to the names of offset .mat files}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error.}

\item{COV_values}{The vector containing the cutoff COV values you would like
to create filtered datasets for.}

\item{COV_mat}{The input filtered reference file generated by the function 'filter.covREF'}

\item{backslip_data}{The list corresponding to combined lateral, vertical and total PDFs.
The default is 'backslip_data'.}
}
\value{
A filtered list of matrices. Each matrix will have a set of PDFs filtered by COV values
as referenced in the COV_mat file.
}
\description{
Combine all offset data from the mat files.
}
