% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat.R
\name{filter_covREF}
\alias{filter_covREF}
\title{Create a reference matrix for filtered COV values}
\usage{
filter_covREF(
  backslip_ref,
  offset_ID,
  offset_type,
  COV_values,
  backslip_data = backslip_data
)
}
\arguments{
\item{backslip_ref}{A reference datatable containing coefficient of variation (COV) values.}

\item{offset_ID}{The column number in the reference datatable 'backslip_ref' corresponding
to the names of offset .mat files}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error.}

\item{COV_values}{The vector containing the cutoff COV values you would like
to create filtered datasets for.}

\item{backslip_data}{The list corresponding to combined lateral, vertical and total PDFs.
The default is 'backslip_data'.}
}
\value{
A datatable with offset_IDs and associated COV values
}
\description{
Create a modified reference table including offset IDs and columns of filtered COV values
}
