% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_import.R
\name{mat2list}
\alias{mat2list}
\title{Combine .mat file data}
\usage{
mat2list(dataDir, dFolder, nameVec)
}
\arguments{
\item{dataDir}{The path to the directory containing a folder of .mat files.}

\item{dFolder}{A folder in your working directory that contains all
the .mat files. Name should be in quotations, i.e. "folder_name".}

\item{nameVec}{A vector with the names of the backslip files.}
}
\value{
A list of three matrices including lateral, vertical and total offset.
}
\description{
Combine all offset data from the mat files.
Make sure each file name starts with a maximum 4-letter/number identifier.
}
