% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_import.R
\name{matrix.mat}
\alias{matrix.mat}
\title{Combine .mat file data}
\usage{
matrix.mat(folder_name, bs_names)
}
\arguments{
\item{folder_name}{A folder in your working directory that contains all
the .mat files. Name should be in quotations, i.e. "folder_name".}

\item{bs_names}{A vector with the names of the backslip files.}
}
\value{
A list of three matrices including lateral, vertical and total offset.
}
\description{
Combine all offset data from the mat files.
Make sure each file name starts with a maximum 4-letter/number identifier.
}
\examples{
backslip_list <- matrix.mat(all_offsets, backslip_vector)
}
