% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{plotMat}
\alias{plotMat}
\title{Plot all backslipped PDFs}
\usage{
plotMat(data_table, offset_type, xmin = -10, xmax = 10, ymin = -1, ymax = 1)
}
\arguments{
\item{data_table}{A single table with x values as the first column and every
subsequent column is a PDF of y values for a single offset site. This table must be
an output from the 'simplify4plot' function.}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error. Also make sure your input offset type corresponds to the
correct datatable.}

\item{xmin}{The minimum x value. Values can be negative or positive but must be less than xmax.}

\item{xmax}{The maximum x value. Values can be negative or positive but must be greater than xmin.}

\item{ymin}{The minimum y value. Values can be negative or positive but must be less than ymax.}

\item{ymax}{The maximum y value. Values can be negative or positive but must be greater than ymin.}
}
\value{
A plot produced by ggplot
}
\description{
Plot PDFs contained in a datatable.
}
