% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{plotMatSum}
\alias{plotMatSum}
\title{Plot summed PDFs}
\usage{
plotMatSum(data_table, offset_type, xmin = -10, xmax = 10)
}
\arguments{
\item{data_table}{A single table with "h_x", "z_x" or "t_x" values as the first column
and a 'y' column of summed PDF values}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error. Also make sure your input offset type corresponds to the
correct datatable.}

\item{xmin}{The minimum x value. Values can be negative or positive but must be less than xmax.}

\item{xmax}{The maximum x value. Values can be negative or positive but must be greater than xmin.}
}
\value{
A plot produced by ggplot
}
\description{
Plot summed PDFs contained in a datatable.
}
