% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{simplify4plot}
\alias{simplify4plot}
\title{Melt backslip data table}
\usage{
simplify4plot(data_table, offset_type)
}
\arguments{
\item{data_table}{A single table with x values as the first column and every
subsequent column is a PDF of y values for a single offset site.}

\item{offset_type}{Input can be "lateral", "vertical" or "total". Any other input
will result in an error.}
}
\value{
A data frame with 3 columns including x values, all backslip names,
and lateral values.
}
\description{
Melt backslip data table into a three column data frame that can be
plotted in ggplot.
}
