% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_import.R
\name{list.mat}
\alias{list.mat}
\title{Import and organize backslip file names into a character vector.}
\usage{
list.mat(dataDir, dFolder)
}
\arguments{
\item{dataDir}{The path to the directory containing a folder of .mat files.}

\item{dFolder}{The name of the folder containing the .mat files}
}
\value{
A character vector listing all the names of .mat files.
}
\description{
Import and organize backslip file names into a character vector.
}
